Отрисовывает оранжевай маркер активного оффера на странице Оффера WIP ошибки типизации
